Increase max_length on majority of model CharFields
The previous limit of 128 characters was too restrictive for
real-world use. The popolo spec doesn't seem to place limits
on the size of attributes, so ideally the Django models wouldn't
either. Using TextField would cause issues when viewing models
in forms, and it's considered to mean something different to
CharField in Django.

There's no mechanism to specify an unlimited number of characters
in a CharField, so I've chosen 4000 as suggested in
https://code.djangoproject.com/ticket/14094