The event's start / end date shouldn't override the membership's
The legislative period's start and end date should only be used for the
membership's start and end date as a default; if the membership has a
start date or end date specified that's likely to be more specific
(e.g. to indicate that someone was only a MP for part of that
legislative period, before someone else took over).

This commit includes a comment explaining that this behaviour isn't
really correct, but is useful until django-popolo has a model for
events.