Allow observers to be notifed of creation / updates during import
This is useful, for example, so that if you want to add your own
identifiers to imported data, you can do that for each object that's
created or updated.

There is a limitation to this at the moment: the setting of parent areas
and parent organizations is done after these callbacks are made, so the
areas and organizations may be missing their parent relationships.