Rather than ignoring an inline area without an ID, raise a ValueError
It seems like dangerous behaviour to just ignore an inline area because
it's missing its 'id' attribute.  Maybe some data in the wild does have
an 'identifier' attribute but no 'id', but let's deal with that if it
comes up - hopefully this exception will make it more likely that we
get a bug report, or spot that that's happening.