Importer: make consistent cases where identifiers is missing and []
If 'identifiers' was missing from a person, the old Identifier objects
associated with a Person would all be preserved, but if instead
identifiers was [] it would remove all Identifier objects associated
with that Person except for the one that preserved the Django ID for
them.

This change makes the behaviour consistent between those two very
similar cases - unknown identifiers are removed when syncing from the
source. (This is what happens for contact details, etc. as well.)